(Remove me: Fix test cases that fail with new paragraph behavior)
Multiline paragraphs are implemented as blocks.
There seems to be a problem with the rendering of blocks not followed by
further input. These test cases have paragraphs as their final input - to
render them they have to end in a newline for now.
(nearly all other test cases do this anyways)